requirements.txt

langchain
langchain-community
langchain-huggingface
langchain-groq
python-dotenv
streamlit
fastapi
uvicorn
autogen